Network initialization and double support
Changes;

- The program now works with doubles, instead of ints.
- You can now give the program a CSV file specifying the structure of a desired neural network, and it will create a new CSV file with randomized values.
- Program has been tested against larger networks, and seems to be working!

+++ Layer.h
@@ -3,7 +3,7 @@
 
 typedef struct {
     int id;
-    int** neurons;
+    double** neurons;
     int neuronCount;
     int weightCount;
 } Layer;
